feat: add user access control with jwt guard

+++ src/user/user.controller.ts
@@ -3,26 +3,28 @@ import {
   Get,
   Post,
   Body,
-  Patch,
   Param,
   Delete,
-  HttpException,
-  Res,
   UseInterceptors,
   Put,
+  UseGuards,
+  Request,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { UserService } from './user.service';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateUserDto } from './dto/update-user.dto';
 import {
   ApiTags,
-  ApiResponse,
   ApiOkResponse,
   ApiBadRequestResponse,
   ApiNotFoundResponse,
+  ApiBearerAuth,
+  ApiUnauthorizedResponse,
 } from '@nestjs/swagger';
 import { UserErrorsInterceptor } from './Interceptor/usererrors.interceptor';
-import { User } from 'src/schemas/user.schema';
+import { Role } from '../constant/role';
+import { JwtAuthGuard } from 'src/auth/jwt-auth.guard';
 @ApiTags('user')
 @Controller('user')
 @UseInterceptors(new UserErrorsInterceptor())
@@ -32,7 +34,7 @@ export class UserController {
     description: 'The user record',
     type: CreateUserDto,
   })
-  @ApiBadRequestResponse({ status: 401, description: '没有权限' })
+  @ApiBadRequestResponse({ status: 401, description: 'Unauthorized' })
   @ApiBadRequestResponse({
     status: 500,
     description:
@@ -41,7 +43,6 @@ export class UserController {
   @Post()
   async create(@Body() createUserDto: CreateUserDto) {
     const result = await this.userService.create(createUserDto);
-    console.log(result);
     return result;
   }
   @ApiOkResponse({
@@ -49,9 +50,23 @@ export class UserController {
     type: CreateUserDto,
     isArray: true,
   })
-  @ApiBadRequestResponse({ status: 401, description: '没有权限' })
+  @ApiBadRequestResponse({ status: 401, description: 'Unauthorized' })
+  @ApiBadRequestResponse({
+    status: 500,
+    description: 'only admin and manager can get user profile',
+  })
   @Get()
-  findAll() {
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth('JWT-auth')
+  findAll(@Request() Req) {
+    if (
+      !Req.user ||
+      (Req.user.role !== Role.admin && Req.user.role !== Role.manager)
+    )
+      throw new UnauthorizedException(
+        undefined,
+        'only admin and manager can get user profile',
+      );
     return this.userService.findAll();
   }
   @ApiOkResponse({
@@ -59,9 +74,23 @@ export class UserController {
     type: CreateUserDto,
   })
   @ApiNotFoundResponse({ description: 'user not found.' })
-  @ApiBadRequestResponse({ status: 401, description: '没有权限' })
+  @ApiBadRequestResponse({
+    status: 500,
+    description: 'only admin and manager can get user profile',
+  })
+  @ApiUnauthorizedResponse({ description: 'Unauthorized' })
   @Get(':username')
-  findOne(@Param('username') username: string) {
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth('JWT-auth')
+  findOne(@Request() Req, @Param('username') username: string) {
+    if (
+      !Req.user ||
+      (Req.user.role !== Role.admin && Req.user.role !== Role.manager)
+    )
+      throw new UnauthorizedException(
+        undefined,
+        'only admin and manager can get user profile',
+      );
     return this.userService.findOne(username);
   }
   @ApiOkResponse({
@@ -69,25 +98,48 @@ export class UserController {
     type: CreateUserDto,
   })
   @ApiNotFoundResponse({ description: 'user not found.' })
-  @ApiBadRequestResponse({ status: 401, description: '没有权限' })
+  @ApiBadRequestResponse({
+    status: 500,
+    description: 'only admin can update user profile',
+  })
+  @ApiUnauthorizedResponse({ description: 'Unauthorized' })
   @ApiBadRequestResponse({
     status: 500,
     description:
       'User validation failed: email: Please fill a valid email address., college: Please fill a valid college., phone: Please fill a valid phone., password: 请输入至少8-16个字符，至少1个大写字母，1个小写字母和1个数字的密码.',
   })
   @Put(':username')
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth('JWT-auth')
   update(
+    @Request() Req,
     @Param('username') username: string,
     @Body() updateUserDto: UpdateUserDto,
   ) {
+    if (!Req.user || Req.user.role !== Role.admin)
+      throw new UnauthorizedException(
+        undefined,
+        'only admin can update user profile',
+      );
     return this.userService.update(username, updateUserDto);
   }
   @ApiOkResponse({ description: 'ok.' })
   @ApiNotFoundResponse({ description: 'user not found.' })
-  @ApiBadRequestResponse({ status: 401, description: '没有权限' })
+  @ApiBadRequestResponse({
+    status: 500,
+    description: 'only admin can delete user profile',
+  })
+  @ApiUnauthorizedResponse({ description: 'Unauthorized' })
   @ApiBadRequestResponse()
   @Delete(':username')
-  remove(@Param('username') username: string) {
+  @UseGuards(JwtAuthGuard)
+  @ApiBearerAuth('JWT-auth')
+  remove(@Request() Req, @Param('username') username: string) {
+    if (!Req.user || Req.user.role !== Role.admin)
+      throw new UnauthorizedException(
+        undefined,
+        'only admin can delete user profile',
+      );
     return this.userService.remove(username);
   }
 }
